Cleanup warnings.

+++ lqiolib/src/srvn_results.cpp
@@ -1,5 +1,5 @@
 /*
- *  $Id: srvn_results.cpp 14000 2020-10-25 12:50:53Z greg $
+ *  $Id: srvn_results.cpp 14112 2020-11-20 20:42:27Z greg $
  *
  *  Created by Martin Mroz on 24/02/09.
  *  Copyright 2009 __MyCompanyName__. All rights reserved.
@@ -54,7 +54,7 @@ void set_general(int v, double c, int i, int pr, int ph)
     LQIO::DOM::__document->setResultValid( v != 0 )
 	.setResultConvergenceValue( c )
 	.setResultIterations( i );
-    if ( ph < 0 || LQIO::DOM::Phase::MAX_PHASE < ph ) throw std::runtime_error( "set_general" );
+    if ( ph < 0 || static_cast<int>(LQIO::DOM::Phase::MAX_PHASE) < ph ) throw std::runtime_error( "set_general" );
     n_phases = ph;
 }
